[docc] - DocC tutorial seems to be working. Symbol references do not work but that actually may not matter - TT
