+++ Sources/SwiftCurrent/Models/Workflow.swift
@@ -5,9 +5,6 @@
 //  Created by Tyler Thompson on 8/26/19.
 //  Copyright © 2021 WWT and Tyler Thompson. All rights reserved.
 //
-//  swiftlint:disable file_length
-
-import Foundation
 
 /**
  A doubly linked list of `FlowRepresentableMetadata`s; used to define a process.
@@ -222,301 +219,5 @@ public final class Workflow<F: FlowRepresentable>: LinkedList<_WorkflowItem> {
     private func setupCallbacks(for node: Element, onFinish: ((AnyWorkflow.PassedArgs) -> Void)?) {
         setupProceedCallbacks(node, onFinish)
         setupBackUpCallbacks(node, onFinish)
-    }
-}
-
-extension Workflow {
-    /**
-     Creates a `Workflow` with a `FlowRepresentable`.
-     - Parameter type: a reference to the first `FlowRepresentable`'s concrete type in the workflow.
-     - Parameter launchStyle: the `LaunchStyle` the `FlowRepresentable` should use while it's part of this workflow.
-     - Parameter flowPersistence: a `FlowPersistence` representing how this item in the workflow should persist.
-     */
-    public convenience init(_ type: F.Type,
-                            launchStyle: LaunchStyle = .default,
-                            flowPersistence: @escaping @autoclosure () -> FlowPersistence = .default) {
-        self.init(FlowRepresentableMetadata(type,
-                                            launchStyle: launchStyle) { _ in flowPersistence() })
-    }
-
-    /**
-     Creates a `Workflow` with a `FlowRepresentable`.
-     - Parameter type: a reference to the first `FlowRepresentable`'s concrete type in the workflow.
-     - Parameter launchStyle: the `LaunchStyle` the `FlowRepresentable` should use while it's part of this workflow.
-     - Parameter flowPersistence: a closure taking in the generic type from the `FlowRepresentable` and returning a `FlowPersistence` representing how this item in the workflow should persist.
-     */
-    public convenience init(_ type: F.Type,
-                            launchStyle: LaunchStyle = .default,
-                            flowPersistence: @escaping (F.WorkflowInput) -> FlowPersistence) {
-        self.init(FlowRepresentableMetadata(type,
-                                            launchStyle: launchStyle) { data in
-            guard case .args(let extracted) = data,
-                  let cast = extracted as? F.WorkflowInput else { return .default }
-            return flowPersistence(cast)
-        })
-    }
-
-    /**
-     Creates a `Workflow` with a `FlowRepresentable`.
-     - Parameter type: a reference to the first `FlowRepresentable`'s concrete type in the workflow.
-     - Parameter launchStyle: the `LaunchStyle` the `FlowRepresentable` should use while it's part of this workflow.
-     - Parameter flowPersistence: a closure returning a `FlowPersistence` representing how this item in the workflow should persist.
-     */
-    public convenience init(_ type: F.Type,
-                            launchStyle: LaunchStyle = .default,
-                            flowPersistence: @escaping () -> FlowPersistence) where F.WorkflowInput == Never {
-        self.init(FlowRepresentableMetadata(type,
-                                            launchStyle: launchStyle) { _ in flowPersistence() })
-    }
-
-    /**
-     Creates a `Workflow` with a `FlowRepresentable`.
-     - Parameter type: a reference to the first `FlowRepresentable`'s concrete type in the workflow.
-     - Parameter launchStyle: the `LaunchStyle` the `FlowRepresentable` should use while it's part of this workflow.
-     - Parameter flowPersistence: a closure returning a `FlowPersistence` representing how this item in the workflow should persist.
-     */
-    public convenience init(_ type: F.Type,
-                            launchStyle: LaunchStyle = .default,
-                            flowPersistence: @escaping (F.WorkflowInput) -> FlowPersistence) where F.WorkflowInput == AnyWorkflow.PassedArgs {
-        self.init(FlowRepresentableMetadata(type,
-                                            launchStyle: launchStyle) { flowPersistence($0) })
-    }
-}
-
-extension Workflow where F.WorkflowOutput == Never {
-    /**
-     Adds an item to the workflow; enforces the `FlowRepresentable.WorkflowOutput` of the previous item matches the `FlowRepresentable.WorkflowInput` of this item.
-     - Parameter type: a reference to the next `FlowRepresentable`'s concrete type in the workflow.
-     - Parameter launchStyle: the `LaunchStyle` the `FlowRepresentable` should use while it's part of this workflow.
-     - Parameter flowPersistence: a `FlowPersistence` representing how this item in the workflow should persist.
-     - Returns: a new workflow with the additional `FlowRepresentable` item.
-     */
-    public func thenProceed<FR: FlowRepresentable>(with type: FR.Type,
-                                                   launchStyle: LaunchStyle = .default,
-                                                   flowPersistence: @escaping @autoclosure () -> FlowPersistence = .default) -> Workflow<FR> where FR.WorkflowInput == Never {
-        let wf = Workflow<FR>(first)
-        wf.append(FlowRepresentableMetadata(type,
-                                            launchStyle: launchStyle) { _ in flowPersistence() })
-        return wf
-    }
-
-    /**
-     Adds an item to the workflow; enforces the `FlowRepresentable.WorkflowOutput` of the previous item matches the `FlowRepresentable.WorkflowInput` of this item.
-     - Parameter type: a reference to the next `FlowRepresentable`'s concrete type in the workflow.
-     - Parameter launchStyle: the `LaunchStyle` the `FlowRepresentable` should use while it's part of this workflow.
-     - Parameter flowPersistence: a `FlowPersistence` representing how this item in the workflow should persist.
-     - Returns: a new workflow with the additional `FlowRepresentable` item.
-     */
-    public func thenProceed<FR: FlowRepresentable>(with type: FR.Type,
-                                                   launchStyle: LaunchStyle = .default,
-                                                   flowPersistence: @escaping () -> FlowPersistence) -> Workflow<FR> where FR.WorkflowInput == Never {
-        let wf = Workflow<FR>(first)
-        wf.append(FlowRepresentableMetadata(type,
-                                            launchStyle: launchStyle) { _ in flowPersistence() })
-        return wf
-    }
-
-    /**
-     Adds an item to the workflow; enforces the `FlowRepresentable.WorkflowOutput` of the previous item matches the `FlowRepresentable.WorkflowInput` of this item.
-     - Parameter type: a reference to the next `FlowRepresentable`'s concrete type in the workflow.
-     - Parameter launchStyle: the `LaunchStyle` the `FlowRepresentable` should use while it's part of this workflow.
-     - Parameter flowPersistence: a `FlowPersistence` representing how this item in the workflow should persist.
-     - Returns: a new workflow with the additional `FlowRepresentable` item.
-     */
-    public func thenProceed<FR: FlowRepresentable>(with type: FR.Type,
-                                                   launchStyle: LaunchStyle = .default,
-                                                   flowPersistence: @escaping @autoclosure () -> FlowPersistence = .default) -> Workflow<FR> where FR.WorkflowInput == AnyWorkflow.PassedArgs {
-        let wf = Workflow<FR>(first)
-        wf.append(FlowRepresentableMetadata(type,
-                                            launchStyle: launchStyle) { _ in flowPersistence() })
-        return wf
-    }
-}
-
-extension Workflow where F.WorkflowOutput == AnyWorkflow.PassedArgs {
-    /**
-     Adds an item to the workflow; enforces the `FlowRepresentable.WorkflowOutput` of the previous item matches the `FlowRepresentable.WorkflowInput` of this item.
-     - Parameter type: a reference to the next `FlowRepresentable`'s concrete type in the workflow.
-     - Parameter launchStyle: the `LaunchStyle` the `FlowRepresentable` should use while it's part of this workflow.
-     - Parameter flowPersistence: a `FlowPersistence` representing how this item in the workflow should persist.
-     - Returns: a new workflow with the additional `FlowRepresentable` item.
-     */
-    public func thenProceed<FR: FlowRepresentable>(with type: FR.Type,
-                                                   launchStyle: LaunchStyle = .default,
-                                                   flowPersistence: @escaping @autoclosure () -> FlowPersistence = .default) -> Workflow<FR> {
-        let wf = Workflow<FR>(first)
-        wf.append(FlowRepresentableMetadata(type,
-                                            launchStyle: launchStyle) { _ in flowPersistence() })
-        return wf
-    }
-
-    /**
-     Adds an item to the workflow; enforces the `FlowRepresentable.WorkflowOutput` of the previous item matches the `FlowRepresentable.WorkflowInput` of this item.
-     - Parameter type: a reference to the next `FlowRepresentable`'s concrete type in the workflow.
-     - Parameter launchStyle: the `LaunchStyle` the `FlowRepresentable` should use while it's part of this workflow.
-     - Parameter flowPersistence: a closure taking in the generic type from the `FlowRepresentable.WorkflowInput` and returning a `FlowPersistence` representing how this item in the workflow should persist.
-     - Returns: a new workflow with the additional `FlowRepresentable` item.
-     */
-    public func thenProceed<FR: FlowRepresentable>(with type: FR.Type,
-                                                   launchStyle: LaunchStyle = .default,
-                                                   flowPersistence: @escaping (FR.WorkflowInput) -> FlowPersistence) -> Workflow<FR> {
-        let wf = Workflow<FR>(first)
-        wf.append(FlowRepresentableMetadata(type,
-                                            launchStyle: launchStyle) { data in
-            guard case.args(let extracted) = data,
-                  let cast = extracted as? FR.WorkflowInput else { return .default }
-            return flowPersistence(cast)
-        })
-        return wf
-    }
-
-    /**
-     Adds an item to the workflow; enforces the `FlowRepresentable.WorkflowOutput` of the previous item matches the `FlowRepresentable.WorkflowInput` of this item.
-     - Parameter type: a reference to the next `FlowRepresentable`'s concrete type in the workflow.
-     - Parameter launchStyle: the `LaunchStyle` the `FlowRepresentable` should use while it's part of this workflow.
-     - Parameter flowPersistence: a closure taking in the generic type from the `FlowRepresentable.WorkflowInput` and returning a `FlowPersistence` representing how this item in the workflow should persist.
-     - Returns: a new workflow with the additional `FlowRepresentable` item.
-     */
-    public func thenProceed<FR: FlowRepresentable>(with type: FR.Type,
-                                                   launchStyle: LaunchStyle = .default,
-                                                   flowPersistence: @escaping (FR.WorkflowInput) -> FlowPersistence) -> Workflow<FR> where FR.WorkflowInput == AnyWorkflow.PassedArgs {
-        let wf = Workflow<FR>(first)
-        wf.append(FlowRepresentableMetadata(type,
-                                            launchStyle: launchStyle) { flowPersistence($0) })
-        return wf
-    }
-
-    /**
-     Adds an item to the workflow; enforces the `FlowRepresentable.WorkflowOutput` of the previous item matches the `FlowRepresentable.WorkflowInput` of this item.
-     - Parameter type: a reference to the next `FlowRepresentable`'s concrete type in the workflow.
-     - Parameter launchStyle: the `LaunchStyle` the `FlowRepresentable` should use while it's part of this workflow.
-     - Parameter flowPersistence: a closure taking in the generic type from the `FlowRepresentable.WorkflowInput` and returning a `FlowPersistence` representing how this item in the workflow should persist.
-     - Returns: a new workflow with the additional `FlowRepresentable` item.
-     */
-    public func thenProceed<FR: FlowRepresentable>(with type: FR.Type,
-                                                   launchStyle: LaunchStyle = .default,
-                                                   flowPersistence: @escaping @autoclosure () -> FlowPersistence = .default) -> Workflow<FR> where FR.WorkflowInput == AnyWorkflow.PassedArgs {
-        let wf = Workflow<FR>(first)
-        wf.append(FlowRepresentableMetadata(type,
-                                            launchStyle: launchStyle) { _ in flowPersistence() })
-        return wf
-    }
-
-    /**
-     Adds an item to the workflow; enforces the `FlowRepresentable.WorkflowOutput` of the previous item matches the `FlowRepresentable.WorkflowInput` of this item.
-     - Parameter type: a reference to the next `FlowRepresentable`'s concrete type in the workflow.
-     - Parameter launchStyle: the `LaunchStyle` the `FlowRepresentable` should use while it's part of this workflow.
-     - Parameter flowPersistence: a closure returning a `FlowPersistence` representing how this item in the workflow should persist.
-     - Returns: a new workflow with the additional `FlowRepresentable` item.
-     */
-    public func thenProceed<FR: FlowRepresentable>(with type: FR.Type,
-                                                   launchStyle: LaunchStyle = .default,
-                                                   flowPersistence: @escaping @autoclosure () -> FlowPersistence = .default) -> Workflow<FR> where FR.WorkflowInput == Never {
-        let wf = Workflow<FR>(first)
-        wf.append(FlowRepresentableMetadata(type,
-                                            launchStyle: launchStyle) { _ in flowPersistence() })
-        return wf
-    }
-
-    /**
-     Adds an item to the workflow; enforces the `FlowRepresentable.WorkflowOutput` of the previous item matches the `FlowRepresentable.WorkflowInput` of this item.
-     - Parameter type: a reference to the next `FlowRepresentable`'s concrete type in the workflow.
-     - Parameter launchStyle: the `LaunchStyle` the `FlowRepresentable` should use while it's part of this workflow.
-     - Parameter flowPersistence: a closure returning a `FlowPersistence` representing how this item in the workflow should persist.
-     - Returns: a new workflow with the additional `FlowRepresentable` item.
-     */
-    public func thenProceed<FR: FlowRepresentable>(with type: FR.Type,
-                                                   launchStyle: LaunchStyle = .default,
-                                                   flowPersistence: @escaping () -> FlowPersistence) -> Workflow<FR> where FR.WorkflowInput == Never {
-        let wf = Workflow<FR>(first)
-        wf.append(FlowRepresentableMetadata(type,
-                                            launchStyle: launchStyle) { _ in flowPersistence() })
-        return wf
-    }
-}
-
-extension Workflow {
-    /**
-     Adds an item to the workflow; enforces the `FlowRepresentable.WorkflowOutput` of the previous item matches the `FlowRepresentable.WorkflowInput` of this item.
-     - Parameter type: a reference to the next `FlowRepresentable`'s concrete type in the workflow.
-     - Parameter launchStyle: the `LaunchStyle` the `FlowRepresentable` should use while it's part of this workflow.
-     - Parameter flowPersistence: a `FlowPersistence` representing how this item in the workflow should persist.
-     - Returns: a new workflow with the additional `FlowRepresentable` item.
-     */
-    public func thenProceed<FR: FlowRepresentable>(with type: FR.Type,
-                                                   launchStyle: LaunchStyle = .default,
-                                                   flowPersistence: @escaping @autoclosure () -> FlowPersistence = .default) -> Workflow<FR> where F.WorkflowOutput == FR.WorkflowInput {
-        let wf = Workflow<FR>(first)
-        wf.append(FlowRepresentableMetadata(type,
-                                            launchStyle: launchStyle) { _ in flowPersistence() })
-        return wf
-    }
-
-    /**
-     Adds an item to the workflow; enforces the `FlowRepresentable.WorkflowOutput` of the previous item matches the `FlowRepresentable.WorkflowInput` of this item.
-     - Parameter type: a reference to the next `FlowRepresentable`'s concrete type in the workflow.
-     - Parameter launchStyle: the `LaunchStyle` the `FlowRepresentable` should use while it's part of this workflow.
-     - Parameter flowPersistence: a closure taking in the generic type from the `FlowRepresentable.WorkflowInput` and returning a `FlowPersistence` representing how this item in the workflow should persist.
-     - Returns: a new workflow with the additional `FlowRepresentable` item.
-     */
-    public func thenProceed<FR: FlowRepresentable>(with type: FR.Type,
-                                                   launchStyle: LaunchStyle = .default,
-                                                   flowPersistence: @escaping (FR.WorkflowInput) -> FlowPersistence) -> Workflow<FR> where F.WorkflowOutput == FR.WorkflowInput {
-        let wf = Workflow<FR>(first)
-        wf.append(FlowRepresentableMetadata(type,
-                                            launchStyle: launchStyle) { data in
-            guard case.args(let extracted) = data,
-                  let cast = extracted as? FR.WorkflowInput else { return .default }
-            return flowPersistence(cast)
-        })
-        return wf
-    }
-
-    /**
-     Adds an item to the workflow; enforces the `FlowRepresentable.WorkflowOutput` of the previous item matches the `FlowRepresentable.WorkflowInput` of this item.
-     - Parameter type: a reference to the next `FlowRepresentable`'s concrete type in the workflow.
-     - Parameter launchStyle: the `LaunchStyle` the `FlowRepresentable` should use while it's part of this workflow.
-     - Parameter flowPersistence: a closure returning a `FlowPersistence` representing how this item in the workflow should persist.
-     - Returns: a new workflow with the additional `FlowRepresentable` item.
-     */
-    public func thenProceed<FR: FlowRepresentable>(with type: FR.Type,
-                                                   launchStyle: LaunchStyle = .default,
-                                                   flowPersistence: @escaping @autoclosure () -> FlowPersistence = .default) -> Workflow<FR> where FR.WorkflowInput == Never {
-        let wf = Workflow<FR>(first)
-        wf.append(FlowRepresentableMetadata(type,
-                                            launchStyle: launchStyle) { _ in flowPersistence() })
-        return wf
-    }
-
-    /**
-     Adds an item to the workflow; enforces the `FlowRepresentable.WorkflowOutput` of the previous item matches the `FlowRepresentable.WorkflowInput` of this item.
-     - Parameter type: a reference to the next `FlowRepresentable`'s concrete type in the workflow.
-     - Parameter launchStyle: the `LaunchStyle` the `FlowRepresentable` should use while it's part of this workflow.
-     - Parameter flowPersistence: a closure returning a `FlowPersistence` representing how this item in the workflow should persist.
-     - Returns: a new workflow with the additional `FlowRepresentable` item.
-     */
-    public func thenProceed<FR: FlowRepresentable>(with type: FR.Type,
-                                                   launchStyle: LaunchStyle = .default,
-                                                   flowPersistence: @escaping @autoclosure () -> FlowPersistence = .default) -> Workflow<FR> where FR.WorkflowInput == AnyWorkflow.PassedArgs {
-        let wf = Workflow<FR>(first)
-        wf.append(FlowRepresentableMetadata(type,
-                                            launchStyle: launchStyle) { _ in flowPersistence() })
-        return wf
-    }
-
-    /**
-     Adds an item to the workflow; enforces the `FlowRepresentable.WorkflowOutput` of the previous item matches the `FlowRepresentable.WorkflowInput` of this item.
-     - Parameter type: a reference to the next `FlowRepresentable`'s concrete type in the workflow.
-     - Parameter launchStyle: the `LaunchStyle` the `FlowRepresentable` should use while it's part of this workflow.
-     - Parameter flowPersistence: a closure returning a `FlowPersistence` representing how this item in the workflow should persist.
-     - Returns: a new workflow with the additional `FlowRepresentable` item.
-     */
-    public func thenProceed<FR: FlowRepresentable>(with type: FR.Type,
-                                                   launchStyle: LaunchStyle = .default,
-                                                   flowPersistence: @escaping (FR.WorkflowInput) -> FlowPersistence) -> Workflow<FR> where FR.WorkflowInput == AnyWorkflow.PassedArgs {
-        let wf = Workflow<FR>(first)
-        wf.append(FlowRepresentableMetadata(type,
-                                            launchStyle: launchStyle) { flowPersistence($0) })
-        return wf
     }
 }
