+++ Sources/SwiftCurrent/SwiftCurrent.docc/Files/LinkedList.swift
@@ -1,0 +1,1 @@
+../../LinkedList/LinkedList.swift
