+++ Sources/SwiftCurrent/SwiftCurrent.docc/Files/LinkedListNode.swift
@@ -1,0 +1,1 @@
+../../LinkedList/LinkedListNode.swift
